requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.17
itsdangerous==2.2.0
sqlmodel==0.0.22
python-dotenv==1.0.1
httpx==0.27.2
openai==1.57.0
anthropic==0.40.0
pdfplumber==0.11.4
pydantic==2.10.3
markdown==3.7
